workflow-dashboard/backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scheduler
apscheduler==3.10.4

# Browser Automation
browser-use==0.1.40
langchain-anthropic==0.1.23
playwright==1.41.2

# Video Analysis
google-generativeai==0.4.0

# Encryption
cryptography==42.0.2

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Notifications
slack-sdk==3.27.1
aiosmtplib==3.0.1

